V5.9.12 au delay

+++ optimize.py
@@ -15,6 +15,8 @@ from njit_funcs import round_dynamic
 from pure_funcs import (
     analyze_fills_slim,
     denumpyize,
+    numpyize,
+    make_compatible,
     get_template_live_config,
     ts_to_date,
     ts_to_date_utc,
@@ -91,13 +93,22 @@ async def main(algorithm=None):
         prog="Optimize multi symbol", description="Optimize passivbot config multi symbol"
     )
     parser.add_argument(
-        "-o",
+        "-oc",
         "--optimize_config",
         type=str,
         required=False,
         dest="optimize_config_path",
         default="configs/optimize/default.hjson",
         help="optimize config hjson file",
+    )
+    parser.add_argument(
+        "-o",
+        "--output_path",
+        type=str,
+        required=False,
+        dest="optimize_output_path",
+        default=None,
+        help="optimize results directory. Defaults to 'results_{algorithm}_{passivbot_mode}/",
     )
     parser.add_argument(
         "-t",
@@ -140,7 +151,7 @@ async def main(algorithm=None):
         required=False,
         dest="passivbot_mode",
         default=None,
-        help="passivbot mode options: [s/static_grid, r/recursive_grid, n/neat_grid, c/clock]",
+        help="passivbot mode options: [r/recursive_grid, n/neat_grid, c/clock]",
     )
     parser.add_argument(
         "-a",
@@ -153,24 +164,70 @@ async def main(algorithm=None):
         help="optimization algorithm options: [p/pso/particle_swarm_optimization, h/hs/harmony_search]",
     )
     parser.add_argument(
+        "-ct",
+        "--clip-threshold",
+        "--clip_threshold",
+        type=float,
+        required=False,
+        dest="clip_threshold",
+        default=None,
+        help="clip_threshold (see opt config for details)",
+    )
+    parser.add_argument(
         "-ser", "--serial", help="optimize symbols singly, not multi opt", action="store_true"
+    )
+    parser.add_argument(
+        "-sm",
+        "--skip_multicoin",
+        "--skip-multicoin",
+        type=str,
+        required=False,
+        dest="skip_multicoin",
+        default=None,
+        help="y/n when using --start dir/, skip multicoin configs (see opt config for details)",
+    )
+    parser.add_argument(
+        "-ss",
+        "--skip_singlecoin",
+        "--skip-singlecoin",
+        type=str,
+        required=False,
+        dest="skip_singlecoin",
+        default=None,
+        help="y/n when using --start dir/, skip single coin configs (see opt config for details)",
+    )
+    parser.add_argument(
+        "-sns",
+        "--skip_non_matching_single_coin",
+        "--skip-non-matching-single-coin",
+        type=str,
+        required=False,
+        dest="skip_non_matching_single_coin",
+        default=None,
+        help="y/n when using --start dir/, skip configs of other symbols (see opt config for details)",
+    )
+    parser.add_argument(
+        "-sms",
+        "--skip_matching_single_coin",
+        "--skip-matching-single-coin",
+        type=str,
+        required=False,
+        dest="skip_matching_single_coin",
+        default=None,
+        help="y/n when using --start dir/, skip configs of same symbol (see opt config for details)",
     )
     parser = add_argparse_args(parser)
     args = parser.parse_args()
-    if args.symbol is None or "," in args.symbol:
-        args.symbol = "BTCUSDT"  # dummy symbol
-    config = await prepare_optimize_config(args)
+    config = prepare_optimize_config(args)
     args = parser.parse_args()
     if algorithm is not None:
         args.algorithm = algorithm
-    if args.symbol is not None:
-        config["symbols"] = args.symbol.split(",")
     if args.serial:
         all_symbols = config["symbols"].copy()
         print(f"running single coin optimizations serially for symbols {all_symbols}")
         for symbol in all_symbols:
-            args.symbol = symbol
-            config = await prepare_optimize_config(args)
+            args.symbols = symbol
+            config = prepare_optimize_config(args)
             await run_opt(args, config)
     else:
         await run_opt(args, config)
@@ -178,37 +235,7 @@ async def main(algorithm=None):
 
 async def run_opt(args, config):
     try:
-        if args.passivbot_mode is not None:
-            if args.passivbot_mode in ["s", "static_grid", "static"]:
-                config["passivbot_mode"] = "static_grid"
-            elif args.passivbot_mode in ["r", "recursive_grid", "recursive"]:
-                config["passivbot_mode"] = "recursive_grid"
-            elif args.passivbot_mode in ["n", "neat_grid", "neat"]:
-                config["passivbot_mode"] = "neat_grid"
-            elif args.passivbot_mode in ["c", "clock"]:
-                config["passivbot_mode"] = "clock"
-            else:
-                raise Exception(f"unknown passivbot mode {args.passivbot_mode}")
-        algorithm = config["algorithm"] if args.algorithm is None else args.algorithm
-        if algorithm in [
-            "p",
-            "pso",
-            "particle_swarm_optimization",
-            "particle-swarm-optimization",
-        ]:
-            config["algorithm"] = "particle_swarm_optimization"
-        elif algorithm in ["h", "hs", "harmony_search", "harmony-search"]:
-            config["algorithm"] = "harmony_search"
-        else:
-            raise Exception(f"unknown optimization algorithm {algorithm}")
-        passivbot_mode = config["passivbot_mode"]
-        assert passivbot_mode in [
-            "recursive_grid",
-            "static_grid",
-            "neat_grid",
-            "clock",
-        ], f"unknown passivbot mode {passivbot_mode}"
-        config.update(get_template_live_config(passivbot_mode))
+        config.update(get_template_live_config(config["passivbot_mode"]))
         config["long"]["backwards_tp"] = config["backwards_tp_long"]
         config["short"]["backwards_tp"] = config["backwards_tp_short"]
         config["exchange"] = load_exchange_key_secret_passphrase(config["user"])[0]
@@ -230,13 +257,13 @@ async def run_opt(args, config):
                 config["short"]["enabled"] = config["do_short"] = False
             else:
                 raise Exception("please specify y/n with kwarg -le/--short")
-        if args.symbol is not None:
-            config["symbols"] = args.symbol.split(",")
+        if args.symbols is not None:
+            config["symbols"] = args.symbols.split(",")
         if args.n_cpus is not None:
             config["n_cpus"] = args.n_cpus
         if args.base_dir is not None:
             config["base_dir"] = args.base_dir
-        if passivbot_mode == "clock":
+        if config["passivbot_mode"] == "clock":
             config["ohlcv"] = True
         print()
         lines = [
@@ -264,17 +291,17 @@ async def run_opt(args, config):
                 cache_dirpath + "market_specific_settings.json"
             ):
                 logging.info(f"fetching data {symbol}")
-                args.symbol = symbol
-                tmp_cfg = await prepare_backtest_config(args)
+                args.symbols = symbol
+                tmp_cfg = prepare_backtest_config(args)
                 if config["ohlcv"]:
                     data = await load_hlc_cache(
                         symbol,
-                        config["inverse"],
-                        config["start_date"],
+                        tmp_cfg["inverse"],
+                        tmp_cfg["start_date"],
                         config["end_date"],
                         base_dir=config["base_dir"],
-                        spot=config["spot"],
-                        exchange=config["exchange"],
+                        spot=tmp_cfg["spot"],
+                        exchange=tmp_cfg["exchange"],
                     )
                     """
                     config["shared_memories"][symbol] = shared_memory.SharedMemory(
@@ -288,62 +315,41 @@ async def run_opt(args, config):
                 else:
                     downloader = Downloader({**config, **tmp_cfg})
                     await downloader.get_sampled_ticks()
-        if config["algorithm"] == "particle_swarm_optimization":
-            from particle_swarm_optimization import ParticleSwarmOptimization
 
-            # prepare starting configs
-            cfgs = []
-            if args.starting_configs is not None:
-                logging.info("preparing starting configs...")
+        # prepare starting configs
+        cfgs = []
+        if args.starting_configs is not None:
+            if os.path.exists(args.starting_configs):
                 if os.path.isdir(args.starting_configs):
-                    for fname in os.listdir(args.starting_configs):
+                    # a directory was passed as starting config
+                    fnames = [f for f in os.listdir(args.starting_configs) if f.endswith(".json")]
+
+                    if "skip_multicoin" in config["starting_configs_filtering_conditions"]:
+                        fnames = [f for f in fnames if "symbols" not in f]
+                    if "skip_singlecoin" in config["starting_configs_filtering_conditions"]:
+                        fnames = [f for f in fnames if "symbols" in f]
+                    if (
+                        "skip_non_matching_single_coin"
+                        in config["starting_configs_filtering_conditions"]
+                    ):
+                        fnames = [f for f in fnames if "symbols" in f or config["symbols"][0] in f]
+                    if "skip_matching_single_coin" in config["starting_configs_filtering_conditions"]:
+                        fnames = [
+                            f for f in fnames if "symbols" in f or config["symbols"][0] not in f
+                        ]
+
+                    for fname in fnames:
                         try:
-                            """
-                            #  uncomment to skip single starting configs with wrong symbol
-                            if not any(x in os.path.join(args.starting_configs, fname) for x in [config["symbols"][0], "symbols"]):
-                                print("skipping", os.path.join(args.starting_configs, fname))
-                                continue
-                            """
                             cfg = load_live_config(os.path.join(args.starting_configs, fname))
                             assert (
-                                determine_passivbot_mode(cfg) == passivbot_mode
+                                determine_passivbot_mode(cfg) == config["passivbot_mode"]
                             ), "wrong passivbot mode"
                             cfgs.append(cfg)
                             logging.info(f"successfully loaded config {fname}")
 
                         except Exception as e:
                             logging.error(f"error loading config {fname}: {e}")
-                elif os.path.exists(args.starting_configs):
-                    try:
-                        cfg = load_live_config(args.starting_configs)
-                        assert determine_passivbot_mode(cfg) == passivbot_mode, "wrong passivbot mode"
-                        cfgs.append(cfg)
-                        logging.info(f"successfully loaded config {args.starting_configs}")
-                    except Exception as e:
-                        logging.error(f"error loading config {args.starting_configs}: {e}")
-            config["starting_configs"] = cfgs
-            particle_swarm_optimization = ParticleSwarmOptimization(config, backtest_wrap)
-            particle_swarm_optimization.run()
-        elif config["algorithm"] == "harmony_search":
-            from harmony_search import HarmonySearch
-
-            # prepare starting configs
-            cfgs = []
-            if args.starting_configs is not None:
-                logging.info("preparing starting configs...")
-                if os.path.isdir(args.starting_configs):
-                    for fname in os.listdir(args.starting_configs):
-                        try:
-                            cfg = load_live_config(os.path.join(args.starting_configs, fname))
-                            assert (
-                                determine_passivbot_mode(cfg) == passivbot_mode
-                            ), "wrong passivbot mode"
-                            cfgs.append(cfg)
-                            logging.info(f"successfully loaded config {fname}")
-
-                        except Exception as e:
-                            logging.error(f"error loading config {fname}: {e}")
-                elif os.path.exists(args.starting_configs):
+                elif args.starting_configs.endswith(".json"):
                     hm_load_failed = True
                     if "hm_" in args.starting_configs:
                         try:
@@ -353,8 +359,9 @@ async def run_opt(args, config):
                                     "long": hm[k]["long"]["config"],
                                     "short": hm[k]["short"]["config"],
                                 }
+                                cfg = sort_dict_keys(numpyize(make_compatible(cfg)))
                                 assert (
-                                    determine_passivbot_mode(cfg) == passivbot_mode
+                                    determine_passivbot_mode(cfg) == config["passivbot_mode"]
                                 ), "wrong passivbot mode in harmony memory"
                                 cfgs.append(cfg)
                             logging.info(f"loaded harmony memory {args.starting_configs}")
@@ -367,18 +374,30 @@ async def run_opt(args, config):
                         try:
                             cfg = load_live_config(args.starting_configs)
                             assert (
-                                determine_passivbot_mode(cfg) == passivbot_mode
+                                determine_passivbot_mode(cfg) == config["passivbot_mode"]
                             ), "wrong passivbot mode"
                             cfgs.append(cfg)
+                            logging.info(f"successfully loaded config {args.starting_configs}")
                         except Exception as e:
                             logging.error(f"error loading config {args.starting_configs}: {e}")
-            config["starting_configs"] = cfgs
+
+        config["starting_configs"] = cfgs
+
+        if config["algorithm"] == "particle_swarm_optimization":
+            from particle_swarm_optimization import ParticleSwarmOptimization
+
+            particle_swarm_optimization = ParticleSwarmOptimization(config, backtest_wrap)
+            particle_swarm_optimization.run()
+        elif config["algorithm"] == "harmony_search":
+            from harmony_search import HarmonySearch
+
             harmony_search = HarmonySearch(config, backtest_wrap)
             harmony_search.run()
     finally:
-        for symbol in config["shared_memories"]:
-            config["shared_memories"][symbol].close()
-            config["shared_memories"][symbol].unlink()
+        if "shared_memories" in config:
+            for symbol in config["shared_memories"]:
+                config["shared_memories"][symbol].close()
+                config["shared_memories"][symbol].unlink()
 
 
 if __name__ == "__main__":
